requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --output-file requirements.txt
altair==6.0.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
antlr4-python3-runtime==4.7.2
    # via latex2sympy2
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==6.2.4
    # via streamlit
certifi==2026.1.4
    # via requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via streamlit
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   pytest
    #   tqdm
gitdb==4.0.12
    # via gitpython
gitpython==3.1.46
    # via streamlit
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core==2.25.2 ; python_full_version >= '3.14'
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-cloud-vision
    #   google-generativeai
google-api-core==2.29.0 ; python_full_version < '3.14'
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-cloud-vision
    #   google-genai
google-cloud-vision-client==2.188.0
    # via google-generativeai
google-auth==2.47.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-vision
    #   google-generativeai
google-auth-httplib2==0.3.0
    # vistreamlit
sympy
google-generativeai
python-dotenv
google-cloud-vision
matplotlib
numpy==3.12.0
    # via math-mentor
google-generativeai==0.8.6
    # via math-mentor
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.76.0
    # via
    #   google-api-core
    #   google-cloud-vision
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
httplib2==0.31.1
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.11
    # via requests
iniconfig==2.3.0
    # via pytest
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonschema==4.26.0
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
latex2sympy2==1.9.1
    # via math-mentor
markupsafe==3.0.3
    # via jinja2
mpmath==1.3.0
    # via sympy
narwhals==2.15.0
    # via altair
numpy==2.4.1
    # via
    #   pandas
    #   pydeck
    #   streamlit
packaging==25.0
    # via
    #   altair
    #   pytest
    #   streamlit
pandas==2.3.3
    # via streamlit
pillow==12.1.0
    # via
    #   math-mentor
    #   streamlit
pluggy==1.6.0
    # via pytest
proto-plus==1.27.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-cloud-vision
protobuf==5.29.5
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-cloud-vision
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
    #   streamlit
pyarrow==22.0.0
    # via streamlit
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.12.5
    # via google-generativeai
pydantic-core==2.41.5
    # via pydantic
pydeck==0.9.1
    # via streamlit
pygments==2.19.2
    # via pytest
pyparsing==3.3.1
    # via httplib2
pytest==9.0.2
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.1
    # via math-mentor
pytz==2025.2
    # via pandas
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   google-api-core
    #   streamlit
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
streamlit==1.53.0
    # via math-mentor
sympy==1.14.0
    # via
    #   latex2sympy2
    #   math-mentor
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.4
    # via streamlit
tqdm==4.67.1
    # via google-generativeai
typing-extensions==4.15.0
    # via
    #   altair
    #   google-generativeai
    #   grpcio
    #   pydantic
    #   pydantic-core
    #   referencing
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via pandas
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.6.3
    # via requests
watchdog==6.0.0 ; sys_platform != 'darwin'
    # via streamlit
